net/ice/base: fix packet type size
Change ptype variable to correctly be 16-bits in ice_prof_map
structure.

Fixes: 51d04e4933e3 ("net/ice/base: add flexible pipeline module")
Cc: stable@dpdk.org

Acked-by: Qiming Yang <qiming.yang@intel.com>

+++ drivers/net/ice/base/ice_flex_type.h
@@ -526,7 +526,7 @@ struct ice_prof_map {
 	u64 context;
 	u8 prof_id;
 	u8 ptype_count;
-	u8 ptype[ICE_MAX_PTYPE_PER_PROFILE];
+	u16 ptype[ICE_MAX_PTYPE_PER_PROFILE];
 };
 
 #define ICE_INVALID_TCAM	0xFFFF
